Reduced locking time
Report: Locking time has now been reduced. Meaning that the main thread isn't being blocked for as long. This gives a visible performance boost due to the fact that the game doesn't appear to freeze anymore.

VAO regeneration happens within a queue. When all regeneration has been completed, the VAO is then added into the main container.

This allows the background thread to do a lot of the work opposed to the main thread for chunk generation.