Removed linked list functionality and return type of 'getNextAvailableObject'
Removed linked list functionality because it wasn't being used in the normal sense. It is meant to be used to garuntee O(1) access but I was using it in O(n) access. Also it allowed for more potential errors such as keeping the linked list connected properly.

Also fixed issue where while loop wouldn't potentially stop execution on release mode due to the assert not being hit

The return type was changed because it couldn't always garuntee to retreieve a valid object that wasn't in use.
Now returning a pointer allows me to check if its null.  If null then don't interact with the object pulled from the object pool and it allows the game to carry on running instead of crashing